requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
argon2-cffi
